Properly refactored with file history.
