+++ xmlapi/corp.go
@@ -1,4 +1,4 @@
-package eveapi
+package xmlapi
 
 const (
 	CorpContactListURL    = "/corp/ContactList.xml.aspx"
